% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_function.r
\name{bindingTF_per_peak}
\alias{bindingTF_per_peak}
\title{Function to generate a TF binding matrix M where i_th row represents
 i_th PWM and j_th column represetns j_th peak with M_i, j=1 if the
  footprint of i_th PWM overlaps j_th peak and 0 otherwise}
\usage{
bindingTF_per_peak(GRpeaks, GRmotif)
}
\arguments{
\item{GRpeaks}{ATAC-seq or DNase-seq peaks in GRanges class where each row
represents the location, read counts, GC content and type
("target","background","no") of the peak}

\item{GRmotif}{Footprint calls from PIQ or CENTIPEDE algorithm in GRanges
class where each row represents the location of each PWM occurrence and its
 purity score from the algorithm. The footprint calls in the forward strand
 and those in the backward strand from the same PWM are not differentiated.}
}
\value{
Returns a TF binding matrix M where i_th row represents i_th PWM
 and j_th column represetns j_th peak with M_i, j=1 if the footprint of
 i_th PWM overlaps j_th peak and 0 otherwise
}
\description{
Function to generate a TF binding matrix M where i_th row represents
 i_th PWM and j_th column represetns j_th peak with M_i, j=1 if the
  footprint of i_th PWM overlaps j_th peak and 0 otherwise
}
\examples{
# Load in the peak file and footprint calls from PIQ or CENTIPEDE algorithm
peak_file <- system.file("extdata", "input_peak_motif.Rdata",
 package = "BiFET")
load(peak_file)
# Generate a TF binding matrix M where i_th row represents
# i_th PWM and j_th column represetns j_th peak with M_i, j=1
# if the footprint of i_th PWM overlaps j_th peak and 0 otherwise
TFbinding.mat <- bindingTF_per_peak(GRpeaks, GRmotif)
TFbinding.mat[1:4, 1:4]
}
\author{
Ahrim Youn
}
