% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_function.r
\name{calculate_enrich_p}
\alias{calculate_enrich_p}
\title{Function to calculate p-value testing if footprints of a TF are
 over-represented in the target set of peaks compared to the background
  set of peaks correcting for the bias arising from the imbalance of
   GC-content and read counts between target and background set}
\usage{
calculate_enrich_p(TFbinding.mat, reads, GCcontent, targetpeak, backgroundpeak)
}
\arguments{
\item{TFbinding.mat}{A TF binding matrix M outputted from the function
bindingTF_per_peak where i_th row represents i_th PWM and j_th column
 represetns j_th peak with M_i,j=1 if the footprint of i_th PWM
  overlaps j_th peak and 0 otherwise}

\item{reads}{A vector of read counts per peak}

\item{GCcontent}{A vector of GC contents per peak}

\item{targetpeak}{A vector of peak IDs belonging to the target set}

\item{backgroundpeak}{A vector of peak IDs belonging to the background set}
}
\value{
Returns a list of parameter alpha_k, p values from BiFET algorithm
 and p values from the hypergeometric test
}
\description{
Function to calculate p-value testing if footprints of a TF are
 over-represented in the target set of peaks compared to the background
  set of peaks correcting for the bias arising from the imbalance of
   GC-content and read counts between target and background set
}
\examples{
# Load in the peak file and footprint calls from PIQ or CENTIPEDE algorithm
peak_file <- system.file("extdata", "input_peak_motif.Rdata",
 package = "BiFET")
load(peak_file)
# Generate a TF binding matrix M where i_th row represents
# i_th PWM and j_th column represetns j_th peak with M_i, j=1
# if the footprint of i_th PWM overlaps j_th peak and 0 otherwise
TFbinding.mat <- bindingTF_per_peak(GRpeaks, GRmotif)
# Other input parameter values for the function calculate_enrich_p
targetpeak <- which(GRpeaks$peaktype == "target")
backgroundpeak <- which(GRpeaks$peaktype == "background")
reads <- GRpeaks$reads
GCcontent <- GRpeaks$GC
result <- calculate_enrich_p(TFbinding.mat, reads,
 GCcontent, targetpeak, backgroundpeak)
 }
}
\seealso{
\code{\link{bindingTF_per_peak}}
}
\author{
Ahrim Youn
}
