% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_function.r
\name{motif_processing}
\alias{motif_processing}
\title{Function to reformat the output matrix of footprint calls from
 PIQ or CENTIPEDE for use in BiFET}
\usage{
motif_processing(motif)
}
\arguments{
\item{motif}{an output matrix of footprint calls from PIQ or CENTIPEDE
algorithm where each row represents the location where each PWM
 was detected, its ID, purity score from the algorithm and the strand: +/-}
}
\value{
Returns a matrix of footprint calls where the suffix ".RC" in the
 id of PWM detected on the reverse strand is removed and redundant counts
  of the same PWM on the same location are removed.
}
\description{
Function to reformat the output matrix of footprint calls from
 PIQ or CENTIPEDE for use in BiFET
}
\examples{
# Load in the output matrix of footprint calls from PIQ or CENTIPEDE
footprints <- system.file("extdata",
"PBMC_PIQ.bed", package = "BiFET")
PBMCmotif <- read.delim(footprints, header = FALSE)
# Process the output matrix of footprint calls from PIQ
# or CENTIPEDE for use in BiFET
PBMCmotif <- motif_processing(PBMCmotif)
head(PBMCmotif)
}
\author{
Ahrim Youn
}
