% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_function.r
\name{bindingTF_per_peak}
\alias{bindingTF_per_peak}
\title{Function to generate a TF binding matrix M where i_th row represents
 i_th PWM and j_th column represetns j_th peak with M_i, j=1 if the
  footprint of i_th PWM overlaps j_th peak and 0 otherwise}
\usage{
bindingTF_per_peak(peaks, motif)
}
\arguments{
\item{peaks}{A matrix of ATAC-seq or DNase-seq peak whose first column
is the chromosome number, the second and third column is the starting
 and ending position of the peak.}

\item{motif}{A matrix of footprint calls outputted from
motif_processing function}
}
\value{
Returns a TF binding matrix M where i_th row represents i_th PWM
 and j_th column represetns j_th peak with M_i, j=1 if the footprint of
 i_th PWM overlaps j_th peak and 0 otherwise
}
\description{
Function to generate a TF binding matrix M where i_th row represents
 i_th PWM and j_th column represetns j_th peak with M_i, j=1 if the
  footprint of i_th PWM overlaps j_th peak and 0 otherwise
}
\examples{
peak_file <- system.file("extdata",
"islet_PBMC_consensus_peak.Rdata", package = "BiFET")
load(peak_file)
# Load in the output matrix of footprint calls from PIQ or CENTIPEDE
footprints <- system.file("extdata",
"PBMC_PIQ.bed", package = "BiFET")
PBMCmotif <- read.delim(footprints, header = FALSE)
# Process the output matrix of footprint calls from PIQ
# or CENTIPEDE for use in BiFET
PBMCmotif <- motif_processing(PBMCmotif)
# Generate a TF binding matrix M where i_th row represents
# i_th PWM and j_th column represetns j_th peak with M_i, j=1
# if the footprint of i_th PWM overlaps j_th peak and 0 otherwise
TFbinding.mat <- bindingTF_per_peak(peaks, PBMCmotif)
head(TFbinding.mat)
}
\seealso{
\code{\link{motif_processing}}
}
\author{
Ahrim Youn
}
